refactor: 계정 설정 페이지 리팩토링
