+++ src/components/form/Label.tsx
@@ -2,7 +2,10 @@ import { ReactNode } from 'react';
 
 function Label({ id, children }: { id: string; children: ReactNode }) {
   return (
-    <label htmlFor={id} className='text-text-primary text-lg font-medium '>
+    <label
+      htmlFor={id}
+      className='text-text-primary text-text-primary dark:text-text-primary-dark text-lg font-medium '
+    >
       {children}
     </label>
   );
